--- backend/network.py
@@ -1,5 +1,0 @@
-import os
-
-
-def get_gateway_ip() -> str:
-    return os.environ['GATEWAY_IP']
