+++ run.py
@@ -2,7 +2,7 @@ from flask import Flask
 from flask_cors import CORS
 
 from backend.logger import logger
-from backend.mode import Mode
+from backend.environment import Environment
 from backend.endpoints.device import device_bp
 from backend.endpoints.master import master_bp
 from backend.endpoints.shared import shared_bp
@@ -16,7 +16,7 @@ def run():
     app.static_folder = f"{app.root_path}/frontend"
 
     app.register_blueprint(shared_bp)
-    if Mode.is_master():
+    if Environment.is_master():
         logger.info("Initializing in master mode")
         app.register_blueprint(master_bp)
     else:
@@ -24,11 +24,11 @@ def run():
         app.register_blueprint(device_bp)
 
     logger.info(
-        f"Running app{' in debug mode' if Mode.debug() else ''}..."
+        f"Running app{' in debug mode' if Environment.debug() else ''}..."
     )
     app.run(
-        debug=Mode.debug(),
-        port=5000 if Mode.is_master() else 5001,
+        debug=Environment.debug(),
+        port=5000 if Environment.is_master() else 5001,
         host="0.0.0.0",
         threaded=True
     )
